+++ src/clients/index.ts
@@ -1,2 +1,2 @@
 export { GLOBAL_CONFIG_STORE_KEYS, AcrossConfigStoreClient } from "./AcrossConfigStoreClient";
-export { HubPoolClient } from "./AcrossHubPoolClient";
+export { HubPoolClient } from "./HubPoolClient";
